requirements.txt
numpy==1.18.0
opencv_python==4.1.2.30
dlib==19.21.1
Pillow==8.2.0
PyQt4==4.11.4
PyQt5==5.15.4
pyserial==3.5
